Refactored collection options to apply options in order
This change allows the caller to perform a series of operations during
collection in an order of the caller's choosing. It also enables the
creation of custom collection options.

In the old scheme it was impossible to filter processes based on user,
for example, because the filtering phase was hard-wired to occur before
the detail collection phase. In the new scheme filtering and detail
collection can be interleaved as necessary.

Exclusion filters were also added as part of this work.